Move getCurrentVariantIndexFromFilters function to Util/Product
